Initial Android.mk (miniperl)
config.h entries should then be linked to the repo variables

Only able to build miniperl binary for the moment.

The android/ folder contains two subfolders (mini, micro)
for the generated configs... To be continued...

The actual config sets perl lib root to /data/local/perl,
That could be modified later for a /system integration.

Signed-off-by: Tanguy Pruvot <tanguy.pruvot@gmail.com>